requirements.txt
boto3==1.40.8
python-dotenv